backend/pyproject.toml
[tool.poetry]
name = "lokalise-integration"
version = "0.1.0"
description = "Lokalise integration with OpenRouter for AI-powered translations"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
Django = "^5.0.1"
python-dotenv = "^1.0.0"
djangorestframework = "^3.14.0"
django-cors-headers = "^4.3.1"
requests = "^2.31.0"
python-lokalise-api = "^3.1.0"
pydantic = "^2.5.3"
celery = "^5.3.6"
redis = "^5.0.1"
python-decouple = "^3.8"
gunicorn = "^21.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.7.0"
isort = "^5.12.0"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3 